Remove dead code.
